# Microservices/MonitoringGateway/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MonitoringGateway",
    "version": "v1"
  },
  "paths": {
    "/api/monitoring-location/get-newest": {
      "post": {
        "tags": [
          "Location"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-location/get-average-h": {
      "post": {
        "tags": [
          "Location"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-location/get-average-day": {
      "post": {
        "tags": [
          "Location"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-location/get-average": {
      "post": {
        "tags": [
          "Location"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AverageDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AverageDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AverageDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-location/get-all": {
      "post": {
        "tags": [
          "Location"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-location/get-timeframe": {
      "post": {
        "tags": [
          "Location"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationTimeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationTimeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LocationTimeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-location/get-traffic-data": {
      "post": {
        "tags": [
          "Location"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRadiusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-subscription/get-newest": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-subscription/get-average-per-hour": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-subscription/get-average-per-day": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-subscription/get-traffic-data": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-time/get-average-h/{year}/{month}/{day}/{hour}": {
      "get": {
        "tags": [
          "Time"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "hour",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-time/get-average-day/{year}/{month}/{day}": {
      "get": {
        "tags": [
          "Time"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/monitoring-time/get-timeframe": {
      "post": {
        "tags": [
          "Time"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeframeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeframeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TimeframeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LocationRadiusDTO": {
        "type": "object",
        "properties": {
          "Latitude": {
            "type": "number",
            "format": "double"
          },
          "Longitude": {
            "type": "number",
            "format": "double"
          },
          "RadiusMeters": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "AverageDTO": {
        "type": "object",
        "properties": {
          "Timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "PerHour": {
            "type": "boolean"
          },
          "LocationInfo": {
            "$ref": "#/components/schemas/LocationRadiusDTO"
          }
        },
        "additionalProperties": false
      },
      "LocationTimeDTO": {
        "type": "object",
        "properties": {
          "Latitude": {
            "type": "number",
            "format": "double"
          },
          "Longitude": {
            "type": "number",
            "format": "double"
          },
          "RadiusMeters": {
            "type": "number",
            "format": "double"
          },
          "ReferenceTime": {
            "type": "string",
            "format": "date-time"
          },
          "Seconds": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TimeframeDTO": {
        "type": "object",
        "properties": {
          "Seconds": {
            "type": "integer",
            "format": "int32"
          },
          "ReferentTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}